+++ tests/data/forms/fixes/form.v.ts
@@ -3,9 +3,9 @@ import { Form } from "../../../../src";
 import FormInterface from "../../../../src/models/FormInterface";
 
 const generateNewPiece = () => ({
-    length: Math.floor(Math.random() * 10),
-    width: Math.floor(Math.random() * 10),
-    height: Math.floor(Math.random() * 10)
+    length: 10,
+    width: 10,
+    height: 10
 });
 
 export default new Form(